+++ src/Providers/Bookmarks/BookmarksController.cs
@@ -1,0 +1,35 @@
+﻿using Windows.UI.Input;
+
+namespace DialToolsForVS
+{
+    internal class BookmarksController : IDialController
+    {
+        public string Moniker => BookmarksControllerProvider.Moniker;
+        public Specificity Specificity => Specificity.Global;
+        public bool CanHandleClick => true;
+        public bool CanHandleRotate => true;
+
+        public void OnClick(RadialControllerButtonClickedEventArgs args, DialEventArgs e)
+        {
+            VsHelpers.ExecuteCommand("Edit.ToggleBookmark");
+            e.Action = "Toggle bookmark";
+            e.Handled = true;
+        }
+
+        public void OnRotate(RotationDirection direction, DialEventArgs e)
+        {
+            if (direction == RotationDirection.Right)
+            {
+                VsHelpers.ExecuteCommand("Edit.NextBookmark");
+                e.Action = "Go to next bookmark";
+            }
+            else
+            {
+                VsHelpers.ExecuteCommand("Edit.PreviousBookmark");
+                e.Action = "Go to previous bookmark";
+            }
+
+            e.Handled = true;
+        }
+    }
+}