+++ src/Providers/Errors/ErrorsControllerProvider.cs
@@ -1,24 +1,22 @@
-﻿using EnvDTE;
-using System.ComponentModel.Composition;
+﻿using System.ComponentModel.Composition;
 using System.IO;
 using System.Reflection;
-using System.Threading.Tasks;
 
 namespace DialToolsForVS
 {
     [Export(typeof(IDialControllerProvider))]
-    internal class DebugControllerProvider : IDialControllerProvider
+    internal class ErrorsControllerProvider : IDialControllerProvider
     {
-        public const string Moniker = "Debug";
+        public const string Moniker = "Errors";
 
         public IDialController TryCreateController(IDialControllerHost host)
         {
             string folder = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-            string iconFilePath = Path.Combine(folder, "Providers\\Debug\\icon.png");
+            string iconFilePath = Path.Combine(folder, "Providers\\Errors\\icon.png");
 
             host.AddMenuItem(Moniker, iconFilePath);
 
-            return new DebugController(host);
+            return new ErrorsController();
         }
     }
 }