+++ CHANGELOG.md
@@ -3,9 +3,8 @@
 - [x] Status bar text update on activation state change
 - [x] Write dial event description in status bar
 - [x] Output Window logging
-- [ ] Add additional menu items
-- [ ] Add bookmark navigation
-- [ ] Add callstack navigation
+- [x] Add bookmark navigation
+- [x] Add debug navigation
 - [ ] Add telemtry
 
 Features that have a checkmark are complete and available for